lien de suppression de catégorie uniquement pour l'admin + redirections

+++ view/forum/listCategories.php
@@ -8,7 +8,11 @@
 foreach($categories as $category ){ ?>
     <p>
         <a href="index.php?ctrl=forum&action=listTopicsByCategory&id=<?= $category->getId() ?>"><?= $category->getName() ?></a>
-        <a href="index.php?ctrl=forum&action=deleteCategory&id=<?= $category->getId() ?>">Supprimer</a>
+        <?php
+        if(App\Session::isAdmin()){
+            ?>
+            <a href="index.php?ctrl=forum&action=deleteCategory&id=<?= $category->getId() ?>">Supprimer</a>
+        <?php } ?>
     </p>
 <?php }
 ?>
